Most up to date for ICE-4. But is only based off the Lab 11 branch - something we did in lab 12 does not jive with the program so I reverted for the time being.
Email sender still does not work properly - even when switching to smpt. Think the differences in both labs have caused some issues with zero direction on how to solve this - which is difficult for someone just learning this framework.

The username change is the only other thing not working. This one might be easier to solve but again, since we are basically just copy and pasting code with little to no explanation it is rather difficult to debug.